+++ cg/cli/set/families.py
@@ -53,7 +53,7 @@ def families(
     identifiers: click.Tuple([str, str]),
 ):
     """Set values on many samples at the same time"""
-    store = context.obj["status"]
+    store = context.obj["status_db"]
     cases = _get_cases(identifiers, store)
 
     if not cases: