+++ cg/cli/workflow/microsalt/base.py
@@ -6,7 +6,8 @@ from pathlib import Path
 
 import click
 
-from cg.apps import hk, lims
+from cg.apps.hk import HousekeeperAPI
+from cg.apps.lims import LimsAPI
 from cg.apps.microsalt.fastq import FastqHandler
 from cg.cli.workflow.microsalt.store import store as store_cmd
 from cg.meta.workflow.microsalt import MicrosaltAnalysisAPI
@@ -22,11 +23,11 @@ LOG = logging.getLogger(__name__)
 @click.pass_context
 def microsalt(context: click.Context, ticket: str, dry_run: bool):
     """Microbial workflow"""
-    context.obj["db"] = Store(context.obj["database"])
-    hk_api = hk.HousekeeperAPI(context.obj)
-    lims_api = lims.LimsAPI(context.obj)
+    context.obj["status_db"] = Store(context.obj["database"])
+    hk_api = HousekeeperAPI(context.obj)
+    lims_api = LimsAPI(context.obj)
     analysis_api = MicrosaltAnalysisAPI(
-        db=context.obj["db"],
+        db=context.obj["status_db"],
         hk_api=hk_api,
         lims_api=lims_api,
         fastq_handler=FastqHandler(context.obj),