+++ tests/cli/workflow/balsamic/conftest.py
@@ -745,6 +745,7 @@ def balsamic_context(
     balsamic_store: Store,
     balsamic_lims: MockLimsAPI,
     balsamic_housekeeper: HousekeeperAPI,
+    trailblazer_api,
 ) -> dict:
     """context to use in cli"""
     balsamic_analysis_api = BalsamicAnalysisAPI(
@@ -753,6 +754,7 @@ def balsamic_context(
         housekeeper_api=balsamic_housekeeper,
         fastq_handler=FastqHandler(server_config),
         lims_api=balsamic_lims,
+        trailblazer_api=trailblazer_api,
     )
     return {
         "BalsamicAnalysisAPI": balsamic_analysis_api,